fix: x86-64 JIT register allocation and patch offset bugs
Three critical bugs fixed in wubu_isa_x86_64.c:

1. **Register mapping mismatch**: emit_mov_reg(&e, 0, X) mapped allocator
   register 0 (r10) instead of rax (x86 register 0). All arithmetic,
   comparison, LOAD, STORE, JZ, JNZ, NEG, NOT operations were using
   wrong registers. Added emit_mov_rax_from_vr/emit_mov_rdi_from_vr/
   emit_mov_vr_from_vr helpers that bypass the VR-to-reg mapping.

2. **Patch rel32 off-by-one**: Jump/branch rel32 calculation used
   'pos + 5' instead of 'pos + 4'. Since pos is the start of the
   rel32 field, the next instruction is at pos+4, not pos+5.
   This caused all forward jumps to land 1 instruction too late.

3. **Gauntlet build fix**: Added missing wubu_isa_vulkan.c,
   wubu_isa_x86_64.c, wubu_isa_spirv.c, wubu_spawn.c to gauntlet
   target link. Removed conflicting jit_stub.c (jit.c from OS
   repo provides the same symbols).

Results:
- Simple JIT tests (arithmetic, comparison, LOAD/STORE, JZ): PASS
- MIR battery: 69/69 PASS
- Gauntlet x86-64: 188/247 pass (up from 0/247 pre-fix)
- Gauntlet other targets: 246/247 pass (unchanged)

+++ test_gauntlet_runner.c
@@ -141,8 +141,8 @@ int main(int argc, char **argv) {
     gauntlet_state_t g;
     gauntlet_init(&g, target_names, N_TARGETS);
 
-    printf("WuBuOS Test Gauntlet — C18 pure · %d ISAs · %u targets%s\n\n",
-           N_TARGETS, g.n_targets, core_only ? " (--core built-in suites)" : "");
+    printf("WuBuOS Test Gauntlet — C18 pure · %ld ISAs · %u targets%s\n\n",
+           (long)N_TARGETS, g.n_targets, core_only ? " (--core built-in suites)" : "");
 
     /* Collect test suites (built-in core first, then the external 19k). */
     const test_entry_t *builtin_suites[] = {
